Release

See merge request ths/smt/carl!14

+++ src/tests/benchmarks/framework/BenchmarkOutput.h
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include <carl/io/streamingOperators.h>
+#include <carl-common/util/streamingOperators.h>
 
 namespace carl {
 
@@ -44,7 +44,7 @@ public:
 			for (const auto& name: names) {
 				if (name.second) {
 					auto it = res.second.find(name.first);
-					if (it != res.second.end()) os << "\t" << (it->second / 1000.0);
+					if (it != res.second.end()) os << "\t" << ((double)it->second / 1000.0);
 					else os << "\tnan";
 				}
 			}
